requirements.txt
opencv-python
pillow
torch
torchvision
gradio
pycocotools
numpy
tqdm
